requirements.txt
flask
twilio
google-genai
reportlab
requests
